Update boston_regression.py
Typecast entire boston dataframe to float

+++ boston_regression.py
@@ -22,6 +22,7 @@ st.write('---')
 
 # Loads the Boston House Price Dataset
 boston = pd.read_csv('Boston_housing_data.csv')
+boston = boston.astype(float)
 X = pd.DataFrame(boston[['CRIM', 'ZN', 'INDUS', 'CHAS', 'NOX', 'RM', 'AGE', 'DIS', 'RAD', 'TAX',
        'PTRATIO', 'B', 'LSTAT']])
 Y = pd.DataFrame(boston['MDEV'])
@@ -33,7 +34,7 @@ st.sidebar.header('Specify Input Parameters')
 
 def user_input_features():
     CRIM = st.sidebar.slider('CRIM', float(X.CRIM.min()), float(X.CRIM.max()), float(X.CRIM.mean()))
-    ZN = st.sidebar.slider('ZN', X.ZN.min(), X.ZN.max(), X.ZN.mean())
+    ZN = st.sidebar.slider('ZN', float(X.ZN.min()), float(X.ZN.max()), float(X.ZN.mean()))
     INDUS = st.sidebar.slider('INDUS', X.INDUS.min(), X.INDUS.max(), X.INDUS.mean())
     CHAS = st.sidebar.slider('CHAS', X.CHAS.min(), X.CHAS.max(), X.CHAS.mean())
     NOX = st.sidebar.slider('NOX', X.NOX.min(), X.NOX.max(), X.NOX.mean())
